requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
mysql-connector-python==8.3.0
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
tzlocal==5.2
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
